Rename the spike project to testProjects/karate-v2
'spike' described how the project started, not what it is: a permanent test
fixture that Karate2UITest depends on. The name also left no room for the v1
fixture needed to cover the other runner path.

karate2-spike/ -> testProjects/karate-v2/, with the packages renamed to match
their roles: spike/ -> sample/ (the passing feature and the feature it calls)
and spike2/ -> broken/ (the deliberately failing one).
