+++ docs/KARATE2.md
@@ -78,7 +78,7 @@ IDE console (phase 2)
 
 | Phase | Contents | Status |
 |---|---|---|
-| 0 | Spike: verify v2 API surface, runner dispatch skeleton, `karate2-spike/` sample project | done |
+| 0 | Spike: verify v2 API surface, runner dispatch skeleton, `testProjects/karate-v2/` sample project | done |
 | 1 | Settings override (AUTO/V1/V2) in `KarateSettingsState` | done |
 | 2 | `KarateV2EventProcessor` (JSON events -> id-based test tree), nesting via callDepth, failure mapping, unit tests | done |
 | 2b | `Karate2UITest` end-to-end integration test (IDE Starter/Driver, screenshots); required reviving Kotlin compilation for `src/integrationTest/kotlin` | this branch |
