--- karate2-spike/settings.gradle.kts
@@ -1,2 +1,0 @@
-// Standalone spike project - intentionally NOT included in the root uppercut build.
-rootProject.name = "karate2-spike"
