+++ testProjects/karate-v2/settings.gradle.kts
@@ -1,0 +1,2 @@
+// Standalone sample project - intentionally NOT included in the root uppercut build.
+rootProject.name = "karate-v2"
